Refactor BufferUtil usage across multiple classes for improved clarity and consistency
