+++ src/test/java/net/openhft/chronicle/bytes/CopyToTest.java
@@ -3,6 +3,7 @@
  */
 package net.openhft.chronicle.bytes;
 
+import net.openhft.chronicle.bytes.util.BufferUtil;
 import net.openhft.chronicle.core.Jvm;
 import org.junit.Test;
 
@@ -19,7 +20,7 @@ public class CopyToTest {
 
         Bytes<?> bytesToTest = Bytes.fromDirect("THIS IS A TEST STRING");
         ByteBuffer copyToDestination = ByteBuffer.allocateDirect(128);
-        copyToDestination.limit((int) bytesToTest.readLimit());
+        BufferUtil.limit(copyToDestination, (int) bytesToTest.readLimit());
         bytesToTest.copyTo(copyToDestination);
         assertEquals("THIS IS A TEST STRING", Bytes.wrapForRead(copyToDestination).toUtf8String());
     }
@@ -28,7 +29,7 @@ public class CopyToTest {
     public void testCopyFromHeapBytesIntoByteBuffer() {
         Bytes<?> bytesToTest = Bytes.from("THIS IS A TEST STRING");
         ByteBuffer copyToDestination = ByteBuffer.allocate(128);
-        copyToDestination.limit((int) bytesToTest.readLimit());
+        BufferUtil.limit(copyToDestination, (int) bytesToTest.readLimit());
         bytesToTest.copyTo(copyToDestination);
         assertEquals("THIS IS A TEST STRING", Bytes.wrapForRead(copyToDestination).toUtf8String());
     }
